Add files via upload
It's all in German and written with Batch. It was my first little "program" and I never finished it. The password is 1234 and If you want that the password activation/deactivation works, you have to put all the files in: "C:\Users\%username%\Desktop\BinData" and create an .lnk to a_BinData.start.lnk and a_BinData2.start.lnk with the names: BinData.lnk and BinData2.lnk. The first .lnk has to be on the desktop and the second inside of: "C:\Users\%username%\Desktop\project_bindata\BinData\pwdState".